Separate historical file load from the rest of the FPDS load (they don't share any functions)
+ Also moved the lookup creation to its own function
